Use "non-buffered" binding
Using binder.setBean, instead of read/writeBean simplifies the code and can be
used in almost all forms. In this example app the bound object is detached JPA
entity, which works as the buffer until we actually call the service method.

The "non-buffered binding" is a must-have when you want to fluently handle class
level bean validators, which are not directly supported by the current Binder.