Native UX
The form UX was by default at the level of basic web app, now upgraded
to the level of native app/rich internet app.

The save button already looked like a primary action, which hints that this
is what you most often click. This kind of buttons in native apps typically
has a enter key configured as a shortcut, like in this form now as well.

Validations are now run lazily while typing. For example when fixing an
error in email field, possibly shown error disappears immediately. Technically,
it disappears when user makes a short pause while typing. The
ValueChaneMode.LAZY provides a great balance between UX and the amount
of client-server communication.

Also, the Save button is not enabled, unless the form actually contains some
changes and is in a valid state.