DRY
Sometime the defaults are just not right. But it is for example not necassiry
to repeat the fact the Grid should configure column width automatically each
time when defining a column. If you see a lot of repetition in your code, it is
usually easy to switch to use loops and/or versions of the API that allow
usage of variable amount of arguments.