Refactoring
After one has built or changed any larger code, it makes sense to give it
a fresh look and see if the same can be written in more readable way.
There are dozens of tricks and styles you can use, but often it is quite
easy to improve the code quality by re-ordering the code, splitting or mergin
to reasonable methods and checking if there are more self documenting
API or naming you could use.

Some things done in this refactoring change:

 * Similar tasks where moved(or merged) to helper methods.
 * Reordered the constructor to first build the layout&configure the presentation, then bind data to grid and form and then finally hook listeners with the actual UI logic
 * Using more descriptive API (deselectAll instead of setValue(null))
 * Better naming of methods and variables